getting closer but convert type lookup is failing

+++ importcsv/models.go
@@ -25,47 +25,47 @@ func MakeModels() ModelFactory {
 
 type Country struct {
    gorm.Model
-   Latitude float32 
-   Have int8 
-   That int8 
-   Longitude int8 
-   In int8 
-   Our int8 
-   Name string 
-   Code string 
-   Other int8 
-   Data int8 
-   We int8 
-   Dont int8 
-   Alias string 
-   Some int8 
-   
+   Latitude float64
+   Have int8
+   That int8
+   Longitude float64
+   In int8
+   Our int8
+   Name string
+   Code string
+   Other int8
+   Data int8
+   We int8
+   Dont int8
+   Alias string
+   Some int8
+
 }
 
 type TestTypes struct {
    gorm.Model
-   Intcol int16 
-   Boolcol bool 
-   Datecol string 
-   Wordcol string 
-   Codecol string 
-   Textcol string 
-   Bigtextcol string 
-   Numbercol float32 
-   
+   Intcol int16
+   Boolcol bool
+   Datecol string
+   Wordcol string
+   Codecol string
+   Textcol string
+   Bigtextcol string
+   Numbercol float32
+
 }
 
 
 func (f ModelFactory) New(name string) interface{} {
 	name = strings.ToLower(name)
 	switch name {
-	
+
 	case "country":
 		return &Country{}
-	
+
 	case "testtypes":
 		return &TestTypes{}
-	
+
 	}
 	return nil
 }
